move routine to pointing module
