--- src/furax/math/__init__.py
@@ -1,5 +1,0 @@
-from .quaternion import get_local_meridian_angle
-
-__all__ = [
-    'get_local_meridian_angle',
-]
